refactor: add import on main.js
