+++ js/data-for-hostel.js
@@ -1,0 +1,50 @@
+export const TITLE_OF_HOTEL = [
+  'Аппарт дворец в Геленджике',
+  'Moscow Inn',
+  'SpaSun',
+  'Красный бермуд',
+  'Сочи-парк',
+  'Отель Прованс',
+  'Holiday inn',
+  'Marriot',
+  'Blue rooms',
+];
+
+export const TYPE_OF_HOTEL = ['palace', 'flat', 'house', 'bungalow'];
+export const TIME_OF_CHECKIN = ['12:00', '13:00','14:00'];
+export const TIME_OF_CHECKOUT = ['12:00','13:00',' 14:00'];
+export const FEATURES_OF_HOTEL = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+export const PHOTO_OF_HOTEL = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export const INTRO = [
+  'Это изумительное ',
+  'Это раскошное ',
+  'Это потрясающее ',
+  'Это бюджетное ',
+  ' Это спокойное ',
+  'Это элитное ',
+  'Это королевское',
+];
+
+export const FEATURES = [
+  'c кондиционером',
+  ' с wi-fi',
+  'c слифтом',
+  'c парковкой',
+  'c душем',
+];
+
+export const TAGLINE = [
+  'доведут тебя к мечтам',
+  'исполнит твои желания',
+  'заставит забыть все невзгоды',
+  'подарит любовь',
+  'создаст комфорт',
+  'наполнит отдых заботой',
+  'вознесет тебя до небес',
+];
